changed kernelUD to same4all=T
after exploring the error in 'grid too small' from the batchUD function, we found that it occurs during the 'getvertices' function when in certain animals with few locations the kernel shape is so large that the 50% contour is beyond the extend of the study grid. For the Wandering Albatross dataset, changing the grid to a larger grid that is the same for all animals fixed this problem; alternatively, one could create a custom grid for each data set and retain same4all=F, but trade-off (speed vs. reliability) needs to be explored.