fixed bootstrap asymptote function
nls throws error when there is only a singular gradient, to avoid the function failing I included a 'try' so that failure of the nls function does not terminate the bootstrap function. Alternative output is instead the mean inclusion value for the maximum sample size, which cannot be higher than any asymptote otherwise estimated by nls.